feat(front): create the fucking swiper when in viewer
